requirements.txt
boto3==1.34.36
git-remote-codecommit==1.17
GitPython==3.1.41
PyGithub==2.3.0
python-gitlab==4.8.0